Split code into several submodules.
